requirements.txt
agbenchmark==0.0.10
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
anthropic==0.3.6
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.0
async-timeout==4.0.3
attr==0.3.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.3
camel-converter==3.0.3
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
dataclasses-json==0.5.14
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
Django==4.2.6
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
faiss-cpu==1.7.4
fastapi==0.99.1
filelock==3.12.4
fire==0.4.0
Flask==3.0.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.37
greenlet==3.0.0
grpcio==1.59.0
grpcio-tools==1.59.0
h11==0.14.0
h2==4.1.0
helicone==1.0.12
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
inquirer==3.1.3
ipython==8.16.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpickle==3.0.2
jsonschema==4.19.1
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lancedb==0.1.16
langchain==0.0.231
langchainplus-sdk==0.0.20
lazy-object-proxy==1.9.0
libcst==1.0.1
litellm==0.1.824
lockfile==0.12.2
loguru==0.6.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
meilisearch==0.21.0
-e git+https://github.com/geekan/metagpt@b9a7a89e252bdb9b782590ef2dab3c314d3e707d#egg=metagpt
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.7
numpy==1.24.3
open-interpreter==0.1.4
openai==0.27.8
openapi-core==0.18.1
openapi-schema-pydantic==1.2.4
openapi-schema-validator==0.6.2
openapi-spec-validator==0.6.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.2
pandas==2.0.3
parse==1.19.1
parso==0.8.3
pathable==0.4.3
peewee==3.16.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
-e git+http://150.109.60.238:3000/tbxy09/gpt-pilot.git@06222986e9fab7dc8623c61126bc3d1edcb8ff3e#egg=pilot
pluggy==1.3.0
portalocker==2.8.2
prance==23.6.21.0
prompt-toolkit==3.0.36
protobuf==4.24.4
psutil==5.9.5
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==13.0.0
pycparser==2.21
pydantic==1.10.13
PyGithub==2.1.1
Pygments==2.16.1
pyhumps==3.8.0
PyJWT==2.8.0
pylance==0.5.10
PyNaCl==1.5.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytz==2023.3.post1
pyvis==0.3.2
PyYAML==6.0
qdrant-client==1.4.0
questionary==2.0.1
ratelimiter==1.2.0.post0
readchar==4.0.5
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.6.0
rpds-py==0.10.4
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.14.0
semantic-kernel==0.3.10.dev0
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
ta==0.10.2
tenacity==8.2.2
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.14.1
tokentrim==0.1.10
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
traitlets==5.11.2
trio==0.22.2
trio-websocket==0.11.1
types-requests==2.31.0.8
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
wcwidth==0.2.8
Werkzeug==3.0.0
wget==3.2
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
yaspin==3.0.1
zipp==3.17.0
